Move Scratch/CopyData class declarations into assembly.h
This allows the use of these classes in other files than just assembly.cc.
